// client/app/shared/models/user.model.ts
export class User {
  _id?: string;
  username?: string;
  email?: string;
  fname?: string;
  lname?: string;
  dob?: string;
  phone_number?: String;
  prifile_pic?: String;
  role?: string;
  country?: String;
  state?: String;
  tags?: String;
  city?: String;
  zip_code?: String;
  provider?: String;
  social_token?: String;
  resetPasswordToken?: string;
}